# docs/api.yaml
openapi: 3.0.1
info:
  title: NAPA project
  description: NAPA trainee assignment
  version: "1"
...
paths:
  /ships:
    post:
      tags: ["ships"]
      description: Adds a new ship to the data base
      operationId: addShip
      requestBody:
       description: Ship details
       required: true
       content:
        application/json:
          schema:
            type: object
            description: The name and speed of the new ship
            properties:
              name:
                $ref: '#/components/schemas/ShipName'
              speed:
                $ref: '#/components/schemas/ShipSpeed'
      responses:
        '201':
          description: Ship created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShipName'
        '400':
          description: Invalid input or parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
    get:
      tags: ["ships"]
      description: Retrieves the names of all the ships
      operationId: getShips
      responses:
        '200':
          description: The list of all ship names.
          content:
            application/json:
              schema:
                description: ship names array
                type: array
                items:
                  $ref: '#/components/schemas/ShipName'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
  /ships/{name}:
    parameters:
     - name: name
       in: path
       required: true  
       description: The name for a ship.
       schema:
        $ref: '#/components/schemas/ShipName'
    put: 
      tags: ["ships"]
      description: Update the name of a specific ship
      operationId: updateShipName
      requestBody:
       description: New ship name
       required: true
       content:
        application/json:
          schema:
           $ref: '#/components/schemas/ShipName'
      responses:
        '204':
          description: Successful update
        '400':
          description: Invalid input or parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '403':
         description: Forbidden. Ship name already exists
         content:
           application/json:
             schema:
               $ref: '#/components/schemas/ErrorMessage'
        '404':
          description: Ship not found.
          content:
           application/json:
             schema:
               $ref: '#/components/schemas/ErrorMessage'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
    delete:
      tags: ["ships"]
      description: Delete a ship
      operationId: deleteShip
      responses:
        '204':
          description: Successful delete
        '404':
          description: Ship not found.
          content:
           application/json:
             schema:
               $ref: '#/components/schemas/ErrorMessage'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
  /ships/{name}/topSpeed: 
    parameters:
     - name: name
       in: path
       required: true  
       description: The name of a ship
       schema:
        $ref: '#/components/schemas/ShipName'
    get: 
      tags: ["ships"]
      description: Retrieves the top speed of a specific ship
      operationId: getShipSpeed
      responses:
       '200':
         description: Ship top speed
         content:
           application/json:
             schema:
               $ref: '#/components/schemas/ShipSpeed'
       '404':
         description: Ship not found.
         content:
          application/json:
           schema:
            $ref: '#/components/schemas/ErrorMessage'
       '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
    put:
      tags: ["ships"]
      description: Update the top speed of a specific ship
      operationId: updateShipSpeed
      requestBody:
       description: New ship speed
       required: true
       content:
        application/json:
          schema:
           $ref: '#/components/schemas/ShipSpeed'
      responses:
        '204':
          description: Successful update
        '400':
          description: Invalid input or parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '404':
          description: Ship not found.
          content:
           application/json:
             schema:
               $ref: '#/components/schemas/ErrorMessage'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
...
components: 
  schemas:
...
    ShipSpeed:
      type: integer
      description: A ships top speed.
      example: 80
    ShipName:
      type: string
      description: The name of a ship.  Also functions as unique identifier.
      pattern: '^.*?$'
      example: Titanic
...
    ErrorMessage:
      type: string
      pattern: '^.*?$'
      example: Error
      description: Error Message